Return Standard TimezoneOffset
I believe there might be people out there who would need to return Standard Timezone Offset. This was a need for me, leading to this addition I'm suggesting here. 
I also added the `timezoneOffset` field to return the timezone so the developer/user can see if the user is on DST or not.